Changement de la structure du vecteur d'impasses
Passage d'une structure en entier à une structure en bit, où le ième bit
représente l'impasse pour la ième ligne.

+++ main.c
@@ -15,7 +15,7 @@ int main(int argc, char* argv[]){
 	int n;
 
 	Element** index = NULL;
-	int* emptyLines = NULL;
+	uint8_t* emptyLines = NULL;
 	double* nabla = NULL;
 	double* delta = NULL;
