+++ read.h
@@ -7,11 +7,13 @@
 
 #ifndef READ_H_
 #define READ_H_
+
 #include <stdio.h>
 #include <stdlib.h>
 #include "structures.h"
+#include "utils.h"
 
-int readFile(char * filename, Element*** pIndex, int* pN, int** pEmptyLines, double** pNabla, double** pDelta);
+int readFile(char * filename, Element*** pIndex, int* pN, uint8_t** pEmptyLines, double** pNabla, double** pDelta);
 
 
 #endif /* READ_H_ */
